+++ CHANGELOG.md
@@ -14,6 +14,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - additional schemata for validation can be declared in SdcProviderComponents and SdcConsumerComponents
 - new method Consumer.restart
 - log a message with level INFO when a context state update is received
+- allow to use full qualified name instead of IP in provider and consumer
 
 ### Fixed